refactor: 수익금 자료형 long으로 변경

+++ src/main/java/blackjack/dto/PlayerGameResult.java
@@ -5,7 +5,7 @@ import blackjack.domain.MatchResult;
 public record PlayerGameResult(
     String nickname,
     MatchResult matchResult,
-    double profit
+    long profit
 ) {
 
 }
